requirements.txt
accelerate==1.6.0
bitsandbytes==0.45.3
datasets==3.3.2
filelock==3.13.1
huggingface-hub==0.29.2
lightning==2.5.0.post0
matplotlib==3.10.1
numpy==2.2.3
psutil==7.0.0
PyQt6==6.8.1
ray==2.43.0
seaborn==0.13.2
sentencepiece==0.2.0
torch==2.6.0 --index-url https://download.pytorch.org/whl/cu124
torchmetrics==1.6.2
tqdm==4.67.1
transformers==4.49.0
wandb==0.19.8
